Add small DI-functionality
inject.ts provides a small type-checked dependency injection framework
with few bells and whistles. Unlike Angular/Nest.js/injection-js and
some others, the type and name of injectables are not tied to classes.
Rather, a names are always strings and a value of any type can be a
depenency. tsc can statically ensure that all dependencies can be
resolved and no run-time errors can appear.

The `inject()` function can be used to both declare interfaces for
abstract services and to define providers for services. For instance,
configuration can be declared with

    interface Config {
       port: number
       env: 'production' | 'development'
    }
    //     name            type
    const Config = inject<Config>()

Other services should preferably agree on the name of injectables, in
this case "Config", but it's not strictly required. A service that
depends on the confiugration is defined as follows:

    interface MyServiceDependencies {
        Config: Config,
        Logger: SomeLogger,
        ...
    }
    //                           injectables (e.g. classes)
    const MyService = inject({ Config, Logger }, services => {
      //      realized values (instances)
      const { Config, Logger } = services

      const myService = ...

      // Return a value or a Promise
      return myService
    })

The first argument specifies the dependencies, using JavaScript's
convenient short-hand notation for `{ Config: Config, ... }`. The
parameter type of the constructor that follows is similar to the
dependency declaration, with values of the injected type.

With one or more services it's possible to string them together,
typically at application startup, using the `assemble()` function:

    const MyAsyncLogger = inject<Logger>(...)
    const myConfig: Config = { ... }
    const assembly = await assemble({
        Config: inject(myConfig),
        MyService,
        MyOtherService,
        Logger: MyAsyncLogger,
        ...
    })

There needs to be a provider for every dependency and since the type
of each provider is rich enough you get a type error if some dependency
cannot be provided. Mutual or circular dependencies are not allowed and
result in a run-time error, however.

As seen above, the `inject()` function can also be called with a single
argument, in which case it creates a constant value provider.

A third parameter to `inject()` defines a cleanup function, which
receives a single parameter - the value returned from the constructor.